refactor(ui): lift format helpers to $lib/format.ts
